requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.2
djangorestframework==3.15.2
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-decouple==3.8
requests==2.32.3
runserver==1.0.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
